+++ test/phone_form_field_test.dart
@@ -2,10 +2,14 @@ import 'package:circle_flags/circle_flags.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:phone_form_field/phone_form_field.dart';
-import 'package:phone_form_field/src/models/phone_number_input.dart';
+import 'package:phone_form_field/src/models/phone_controller.dart';
+import 'package:phone_form_field/src/widgets/country_picker/country_selector.dart';
+import 'package:phone_form_field/src/widgets/flag_dial_code_chip.dart';
 
 void main() {
   group('PhoneFormField', () {
+    final formKey = GlobalKey<FormState>();
+    final phoneKey = GlobalKey<FormFieldState>();
     Widget getWidget({
       Function(PhoneNumber?)? onChanged,
       Function(PhoneNumber?)? onSaved,
@@ -17,14 +21,18 @@ void main() {
     }) =>
         MaterialApp(
           home: Scaffold(
-            body: PhoneFormField(
-              initialValue: initialValue,
-              onChanged: onChanged,
-              onSaved: onSaved,
-              showFlagInInput: showFlagInInput,
-              controller: controller,
-              defaultCountry: defaultCountry,
-              phoneNumberType: phoneNumberType,
+            body: Form(
+              key: formKey,
+              child: PhoneFormField(
+                key: phoneKey,
+                initialValue: initialValue,
+                onChanged: onChanged,
+                onSaved: onSaved,
+                showFlagInInput: showFlagInInput,
+                controller: controller,
+                defaultCountry: defaultCountry,
+                phoneNumberType: phoneNumberType,
+              ),
             ),
           ),
         );
@@ -46,13 +54,13 @@ void main() {
 
     testWidgets('Should have a default country', (tester) async {
       await tester.pumpWidget(getWidget(defaultCountry: 'FR'));
-      expect(find.text('+ 33'), findsOneWidget);
+      expect(find.text('+ 33'), findsWidgets);
     });
 
     testWidgets('Should display initial value', (tester) async {
       await tester.pumpWidget(getWidget(
           initialValue: PhoneParser().parseWithIsoCode('FR', '478787827')));
-      expect(find.text('+ 33'), findsOneWidget);
+      expect(find.text('+ 33'), findsWidgets);
       expect(find.text('478787827'), findsOneWidget);
     });
 
@@ -83,13 +91,13 @@ void main() {
       await tester
           .pumpWidget(getWidget(controller: controller, defaultCountry: 'US'));
       controller.value = PhoneParser().parseWithIsoCode('FR', '488997722');
-      await tester.pump();
-      expect(find.text('+ 33'), findsOneWidget);
+      await tester.pumpAndSettle();
+      expect(find.text('+ 33'), findsWidgets);
       expect(find.text('488997722'), findsOneWidget);
     });
 
     testWidgets(
-        'Should change value of dial code chip when full number written',
+        'Should change value of dial code chip when full number copy pasted',
         (tester) async {
       final controller = PhoneController(null);
       // ignore: unused_local_variable
@@ -140,25 +148,33 @@ void main() {
         saved = true;
         phoneNumber = p;
       };
-      await tester.pumpWidget(
-        getWidget(
-          initialValue: phoneNumber,
-          onSaved: onSaved,
-        ),
-      );
+      await tester.pumpWidget(getWidget(
+        initialValue: phoneNumber,
+        onSaved: onSaved,
+      ));
       final foundTextField = find.byType(TextFormField);
-      final foundBase = find.byType(BasePhoneFormField);
-      final txtField = tester.widget<TextFormField>(foundTextField);
-      final base = tester.widget<BasePhoneFormField>(foundBase);
-      expect(base.onSaved != null, isTrue);
-      // non digits should not work
       await tester.enterText(foundTextField, '479281938');
-      base.onSaved
-          ?.call(PhoneNumberInput(isoCode: 'FR', national: '499887766'));
+      await tester.pumpAndSettle();
+      formKey.currentState?.save();
       await tester.pumpAndSettle();
       expect(saved, isTrue);
       expect(phoneNumber,
-          equals(PhoneParser().parseWithIsoCode('FR', '499887766')));
+          equals(PhoneParser().parseWithIsoCode('FR', '479281938')));
+    });
+
+    testWidgets('Should reset', (tester) async {
+      PhoneNumber? phoneNumber =
+          PhoneParser().parseWithIsoCode('FR', 'national');
+
+      await tester.pumpWidget(getWidget(initialValue: phoneNumber));
+      await tester.pumpAndSettle();
+      final national = '123456';
+      final foundTextField = find.byType(TextFormField);
+      await tester.enterText(foundTextField, national);
+      expect(find.text(national), findsOneWidget);
+      formKey.currentState?.reset();
+      await tester.pumpAndSettle();
+      expect(find.text(national), findsNothing);
     });
 
     testWidgets('Should tell when a phone number is not valid', (tester) async {
@@ -202,7 +218,7 @@ void main() {
 
     testWidgets('Should show / not show flag', (tester) async {
       await tester.pumpWidget(getWidget(showFlagInInput: true));
-      expect(find.byType(CircleFlag), findsOneWidget);
+      expect(find.byType(CircleFlag), findsWidgets);
       await tester.pumpWidget(getWidget(showFlagInInput: false));
       expect(find.byType(CircleFlag), findsNothing);
     });